Fix type instability in `_dot_internal` (#630)
* refactor: improve type stability of `_dot_internal`

* refactor: ensure type stability for _dot_internal in ext

* refactor: ensure type stability for _dot_internal

* refactor: fix boxing in _dot_internal

* refactor: copy boxing fix

* chore: bump version

* chore: bump version

+++ ext/MooncakeDynamicExpressionsExt.jl
@@ -241,7 +241,7 @@ Mooncake._dot_internal(c::Mooncake.MaybeCache, t::TangentNode, s::NoTangent) = 0
         deg = t.degree
         res = if deg == 0
             if t.constant && s.constant
-                Mooncake._dot_internal(c, t.val, s.val)
+                Mooncake._dot_internal(c, t.val, s.val)::Float64
             else
                 0.0
             end
@@ -252,7 +252,9 @@ Mooncake._dot_internal(c::Mooncake.MaybeCache, t::TangentNode, s::NoTangent) = 0
                 i -> Base.Cartesian.@ncall(
                     i,
                     +,
-                    j -> Mooncake._dot_internal(c, get_child(t, j), get_child(s, j))
+                    j -> Mooncake._dot_internal(
+                        c, get_child(t, j), get_child(s, j)
+                    )::Float64
                 )
             )
         end
